recommendation-engine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
opensearch-py==2.4.0
redis==5.0.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.15.0
scipy==1.11.1
joblib==1.3.1
mlflow==2.5.0
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
aiofiles==23.1.0
implicit==0.7.2
surprise==1.1.3
lightgb==4.0.0
xgboost==1.7.6
optuna==3.3.0 